+++ src/main/java/com/squadb/workassistantapi/book/domain/NotRentableException.java
@@ -1,4 +1,4 @@
-package com.squadb.workassistantapi.domain.exceptions;
+package com.squadb.workassistantapi.book.domain;
 
 public class NotRentableException extends RuntimeException {
     public NotRentableException(String message) {
