+++ src/main/java/com/squadb/workassistantapi/rental/application/RentalService.java
@@ -2,20 +2,18 @@ package com.squadb.workassistantapi.rental.application;
 
 import com.squadb.workassistantapi.book.application.BookService;
 import com.squadb.workassistantapi.book.domain.Book;
+import com.squadb.workassistantapi.book.domain.NotRentableException;
 import com.squadb.workassistantapi.member.application.MemberService;
 import com.squadb.workassistantapi.member.domain.Member;
 import com.squadb.workassistantapi.member.dto.LoginMember;
 import com.squadb.workassistantapi.rental.domain.Rental;
 import com.squadb.workassistantapi.rental.domain.RentalRepository;
 import com.squadb.workassistantapi.rental.domain.RentalStatus;
-import com.squadb.workassistantapi.domain.*;
-import com.squadb.workassistantapi.domain.exceptions.NotRentableException;
 import com.squadb.workassistantapi.reservation.domain.Reservation;
 import com.squadb.workassistantapi.reservation.domain.ReservationRepository;
 import com.squadb.workassistantapi.reservation.domain.ReservationStatus;
 import com.squadb.workassistantapi.reservation.domain.ReservationValidator;
 import com.squadb.workassistantapi.reservation.dto.ReservationSearchDto;
-import com.squadb.workassistantapi.web.controller.dto.LoginMember;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
