+++ src/main/java/com/squadb/workassistantapi/reservation/presentation/ReservationController.java
@@ -1,13 +1,13 @@
 package com.squadb.workassistantapi.reservation.presentation;
 
 import com.fasterxml.jackson.databind.exc.ValueInstantiationException;
+import com.squadb.workassistantapi.member.dto.LoginMember;
+import com.squadb.workassistantapi.member.infrastructure.config.CheckPermission;
+import com.squadb.workassistantapi.member.infrastructure.config.CurrentLoginMember;
 import com.squadb.workassistantapi.reservation.application.ReservationService;
 import com.squadb.workassistantapi.reservation.domain.Reservation;
 import com.squadb.workassistantapi.reservation.domain.ReservationException;
 import com.squadb.workassistantapi.reservation.dto.*;
-import com.squadb.workassistantapi.web.config.auth.CurrentLoginMember;
-import com.squadb.workassistantapi.web.controller.dto.LoginMember;
-import com.squadb.workassistantapi.web.interceptor.CheckPermission;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
