+++ src/test/java/com/squadb/workassistantapi/book/presentation/BookApiControllerTest.java
@@ -6,6 +6,7 @@ import com.squadb.workassistantapi.book.dto.BookRegisterDto;
 import com.squadb.workassistantapi.book.infrastructure.BookSearchAgent;
 import com.squadb.workassistantapi.member.domain.MemberType;
 import com.squadb.workassistantapi.member.dto.LoginMember;
+import com.squadb.workassistantapi.member.presentation.MemberController;
 import com.squadb.workassistantapi.rental.domain.NoAuthorizationException;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
