+++ src/test/java/com/squadb/workassistantapi/reservation/domain/ReservationTest.java
@@ -1,7 +1,7 @@
 package com.squadb.workassistantapi.reservation.domain;
 
-import com.squadb.workassistantapi.domain.Book;
-import com.squadb.workassistantapi.domain.Member;
+import com.squadb.workassistantapi.book.domain.Book;
+import com.squadb.workassistantapi.member.domain.Member;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
