+++ src/test/java/com/squadb/workassistantapi/reservation/presentation/ReservationControllerTest.java
@@ -2,14 +2,18 @@ package com.squadb.workassistantapi.reservation.presentation;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.squadb.workassistantapi.domain.*;
+import com.squadb.workassistantapi.book.domain.Book;
+import com.squadb.workassistantapi.book.domain.Isbn;
+import com.squadb.workassistantapi.book.domain.StockQuantity;
+import com.squadb.workassistantapi.member.domain.Member;
+import com.squadb.workassistantapi.member.domain.MemberType;
+import com.squadb.workassistantapi.member.dto.LoginMember;
+import com.squadb.workassistantapi.member.presentation.MemberController;
 import com.squadb.workassistantapi.reservation.domain.Reservation;
 import com.squadb.workassistantapi.reservation.domain.ReservationRepository;
 import com.squadb.workassistantapi.reservation.domain.ReservationStatus;
 import com.squadb.workassistantapi.reservation.domain.ReservationValidator;
 import com.squadb.workassistantapi.reservation.dto.ReservationRequestDto;
-import com.squadb.workassistantapi.web.controller.MemberController;
-import com.squadb.workassistantapi.web.controller.dto.LoginMember;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -24,7 +28,7 @@ import javax.persistence.EntityManager;
 import java.time.LocalDateTime;
 import java.util.List;
 
-import static com.squadb.workassistantapi.domain.MemberType.ADMIN;
+import static com.squadb.workassistantapi.member.domain.MemberType.ADMIN;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.mock;
 import static org.springframework.http.MediaType.APPLICATION_JSON;
